requirements.txt
asgiref==3.8.1
Django==5.1.4
djangorestframework==3.15.2
psycopg==3.2.3
python-dotenv==1.0.1
sqlparse==0.5.2
typing_extensions==4.12.2
pillow==11.0.0
psycopg2-binary==2.9.10
django-filter==24.3
djangorestframework-simplejwt==5.3.1
django-celery-beat==2.7.0
drf-yasg==1.21.8
stripe==11.3.0
celery==5.4.0
redis==5.2.1


